CartesianIndex version of scalar eqs
Rewrote the scalar transport equations using CartesianIndex and some new 
helper functions like \delta(a,I) which is a step in direction a. 

This lets all the difference and interpolation formulas be written 
independant of coordinate direction. 

The new code is a bit faster, shorter, and can be generalized to 3D with 
minimal effort.