Clean up example and plot routines
Added shift argument to flood(). The vorticity ω₃ lives in the corner of 
a cell in 2D, so it needs to be shifted down and right half a unit. 
Pressure is cell-centered, so no shift. velocity is face centered, 
etc...

Also renamed `addbody` routine since I might need `body` to hold the 
geometry coordinates, etc soon.