New function sim_remeasure!
This function remeasures an updated SDF to update the simulation 
coefficients. 

The SDF can be updated by changing a parameter value (such as using 
Pluto @bind) or just updating the time variable SDF with the current 
sim_time. 

The Flow body fields are updated (mu0,mu1) as well as the Poisson 
matrix.