Added coordinate mapping to AutoBody to enable dynamics
Changed the AutoBody struct to also carry an optional coordinate 
mapping, with default
map = (x,t)->x. It is composed automatically with the sdf.

Since the map returns a vector, its time-derivative is a vector 
velocity, and is the implicitly defined body's velocity when sdf=0.

Also added a correction to the measure function which approximately 
compensates for using an inexact sdf.

Added two new tests of these two new capabilities.