Enabled dynamic simulations
Added a function WaterLily.time which gives the _unscaled_ time. Use 
this in the (renamed) measure!(sim) function and sim_step!(sim) as this 
is the time used by AD.

Updated TwoD_block.jl for a dynamics plate and it works! The MultiLevel 
solver is struggling with this test, but it runs and the results look 
good.