Added continuous van Leer limiter
W Rider suggested that a continuous limiter avoids issues with 
regression tests and improves adjoints. I've added the symmetric van 
Leer limiter, which is smooth but more dissipative than QUICK. I've 
added an option in the ϕu function to specify the limiter, with quick 
being the default. 

The van Leer limiter benchmarks faster than QUICK. While the limiter 
itself is slightly faster, the real speed up comes from fewer Poisson 
iterations needed. This is likely because of the increased dissipation. 
This may make it a better default, but it needs more testing first.

Note that the "edges" of the monotonic interpolation region are still 
not smooth. All limiters (to my knowledge) have a hard switch to 
O1-upwind when the field is not monotonic. I'm not sure if this is an 
issue for adjoints.