Add body velocity divergence σᵥ
The source term to the Poisson equation needs a correction when the body 
velocity is nonzero.
This is why the previous dynamic simulations were slow to converge and 
didn't converge at all in some cases. 

The correction is computed and stored in measure!(Flow,body). This adds 
one more scalar to the memory footprint, but it would be a waste to 
compute this zero correction twice every time step even when there 
wasn't a moving body.