Version 0.2.3
Added more advanced Pluto notebook example of a swimming shark. This more complex geometry required some upgrades to the AutoBody functions.

This distance function for the geometry is not exact, and I added a correction to `measure!(Flow,AutoBody)` to improve the sdf results near the interface. The sdf function is also expensive to evaluate (nested splines), and I refactored `measure!(Flow,AutoBody)` and added the `apply_sdf!()` function to minimize the number of function evaluation to immerse the body in the flow. Benchmark for measure!() reduced from 18.8ms for v0.2.2 to 3.36ms for v0.2.3.