Corrected error in AutoBody velocity
Given a mapping `ξ=m(x,t)` we need the
material derivative of the map to be zero on the body, ie
`Dm/Dt=0 → ṁ + (dm/dx)ẋ = 0 ∴  ẋ =-(dm/dx)\ṁ`
where `dm/dx` is the Jacobian. I had neglected the Jacobian part and
this is why the flapping case gave the pressure solver issues before.