Clean-up the AutoBody set code
Added a few tests for the AutoBody set code. These found a bug in `intersectBodies` and I don't know how to extend the fix to `morphBodies`, so I've removed it for now.

I also moved that code into AutoBody.jl and exported the overload `+-` for use in setting up simulations.