Update benchmark and @loop
Cleaned up the benchmarking code to be use a 2D and 3D sphere test, as well as adding profiling.

2D is running fine, but 3D has some array allocation issue that I need to track down and fix.

The code is currently spending around 80% of its time on loops created with the @loop macro. I added Polyester.@batch to this macro which does a nice job of maintaining approximately the same optimizations on 1 thread, but improving performance on a multi-core machine. This required changing a few closures in MultiLevelPoisson, but otherwise, it's a 1-line improvement.