Memory leak fix
The memory leak was (of course) from the small closure changes I made separate closure functions, and this let Polyester work and killed the allocations. I've also added a function DISABLE_PUSH() which I can use to quickly stop pushing to the CFL and Poisson count arrays. This lets me ensure the allocations are zero when running on a single thread. When running multi-threaded, there are allocations created for each loop - currently this is a O(100) allocations and only O(10 KiB).