Fixed thread conflict
The flux calculation in conv_diff! contributes to the forcing on the cells on either side of the face. I was doing this all in one loop, but this introduces a multi-threading conflict, so I was getting random forcing. I've passed in an array to calculate and store the flux, and then apply it to the force in two loops to avoid the conflicts.