Extend @loop macro to deal with structs
Added a sketch of the Flow struct (using a mapping to convert to CuArrays, which might be a nice way to do it).

The previous version of the @loop macro failed when given an expression with composite values like
```
@loop flow.div[I] = div_operator(I,flow.u)
```
which is a problem since there are TONS of these in the code-base.
Problem 1: `:(flow.u)` is an Expr not a Symbol, so `grab!()` broke this up into `:flow, :u`, which isn't what we want to pass to the function. This is solved by simply aborting the recursion when `head==:.`.
Problem 2: `:(flow.u)` is not a valid function argument, meaning you can not write
```
function f(flow.div,flow.u) # invalid arguments
   I = @index()
   flow.div[I] = div_operator(I,flow.u)
end
```
Instead you need to replace these with the composite value itself *including in the expression* ie
```
function f(div,u) # rename arguments
   I = @index()
   div[I] = div_operator(I,u) # replace arguments in expression
end
f(flow.div,flow.u) # call with full composite values
```
This was pretty easy, with some help from a discourse post. Now it's running and tested to match the hand-written kernel.