Starting to integrate changes and test
I have commented out most of the WaterLily.jl and test\runtests.jl files. The plan is to roll out the changes to the code base one file at a time, and add the tests as we go.

Currently tested most of the serial code in util, including OA and the new version of inside().