Replaced GS! with SOR! and Jacobi!
Both GS! and SOR! method require progressively scanning through the field so the updates that have already been calculated improve the next update. This massively improves the convergence properties, but it means these approaches are not trivially parallelized. In a block decomposition, you can just apply these within each block and then update the block boundaries. This saves some part of the performance improvement. I don't know how to translate this to a multi-threaded approach, and the way I was doing it before results in race-conditions and random convergence rates when multi-threading.
A Jacobi smoother is a simple multiplication, so it's trivially parallel, but the convergence is really slow.
I've updated solver!(::Poisson) to use scalar SOR! on CPU and Jacobi!(it=20) otherwise. This passes the previous tests and give repeatable results. I've used Jacobi! within solver!(::MultilevelPoisson) but this is still not converging yet.