Indexing update for OffsetArrays and other fixes
The up()/down() functions map one set of indices to another, and this depends on the axes. I have hard coded Origin(0) for all the arrays in MultiLevel Poisson, and added tests that the restriction is completing properly for a non-trivial test case. Everything is now converging as expected.

Also fixed the update!() functions, and added a test.

Note solver!() is using SOR! at the moment, so running in serial.