Starting roll-back of OffsetArrays
We hit a roadblock when applying the current approach to Flow, and suddenly realized that instead of using OffsetArrays, we could just offset the index within the kernel functions! This is much more general and clean. However, it means we need to roll back all the OffsetArray crap.

Started that process by going through util.f90 and it's tests.