Update Flow
Cosmetic change to Flow struct (getting rid of conflicting definitions of V) and expanded the sum in flux_out.

Streamlined the benchmark test in CUDAEnv\Flow including a serial baseline. The @kernel version is quite a bit faster on my laptop because of conv_diff, but the BC! is actually much slower. The GPU version sees up to 250x speed-up. The solver!() is by far the slowest GPU routine now.