Testing pcg! smoothing
The PCG smoother runs on all backends, and it's much faster than SOR! even for serial runs. This must be because of @simd (which can't be used with SOR!). However, it's doesn't speed up at all using @kernel!
I _think_ this is mostly due to using \cdot for all the inner products. These aren't multi-threaded and aren't any faster on CuArrays.
The allocations are also a problem. I needed to add a maxlevel limit for MultilevelPoisson to avoid the @kernel versions slowing down the solver.