Updated @loop and benchmarks
I added @fastmath @inbounds to the expression within the @loop, which plays much nicer with the @kernel script and speeds up the loops significantly.

I used this to streamline pcg! a little since it's performance critical. HOWEVER, I was timing everything on the GPU incorrectly. The speed-up on the GPU is the same as multi-threaded except for really big simulations. On the plus side - this speed-up is real, and so I can measure it on VCycle!, which is now 9x faster than serial for a big 3D simulation.