Updated mom_step.jl benchmark and addeed new tgv.jl benchmark.
The mom_step.jl benchmark was to trivial (flow was mostly 0-filled arrays) so the compiler was probably doing some tricks to skip operations.
Adding the TGV initialisation provides higher speedups, as it is a non-trivial flow.

The new tgv.jl benchmark quantifies the full solver execution for the TGV test case during 0.1 CTU.
The speedups are aligned to what we observe when breaking down mom_step.