Update ThreeD_donut.jl
Simulating on GPU and running visualization on CPU. It's working and is fast enough to run live for a demo. However, the omega_theta metric is allocating like crazy and won't run on the GPU. This is the slowest part of the time loop currently! Fixing it could increase the speed up to 40%.