Updated Metrics.jl
Most of Metrics.jl didn't actually run on the GPU, but now they all do. The GPU is extremely type sensitive and also doesn't play nice with many of the LinearAlgebra functions. I've written a few helper functions to consistently use tuples or SA in the metrics and changed the metrics.jl tests to attempt all of them on the CPU kernel.