Added correct_div! function to account for small changes in body volume
Incompressible solvers don't like it when they are given BCs that aren't globally volume conserving. Added routine that calculates the global divergence in the prescribed velocity and corrects it to zero by subtracting a (hopefully tiny) amount at every cell.
Note that this is a hack and not very physical! But it's a tiny issue for the vast majority of body motions.