Added AMDGPU package
The solver can be run on a AMD GPU using Julia 1.9.0. This has been tested at the BSC AMD cluster with a TGV case.
To run on AMD, simply change the array type from Array to ROCArray.
Tests now also consider AMD devices if available.
The TOML.jl package had to be explicitely added too, even if is part of the standard library - unsure why.