Working exit prototype
The main problem with the previous version were the `a.u .= 0` and `a.u ./= 2` lines in `mom_step!` which kept modifying the exit plane values. I changed these lines to only update the values inside the domain.

I added an `flow.exit` to the struct and initialization functions and use this to trigger the convection exit behavior.

The `apply!` function had the opposite problem - it WASN'T filling in the boundary value. Now it does.

It also seemed weird to me that the convection exit value in Lotus was based on the old value at the boundary but the u* value (before projection) upstream. I'm not sure it matters much, but it seems more consistent to just use the old values in both places, so that's what it does now.

Finally, I added a Lamp vortex dipole test case, which is nice since it should leave the exit without changing size or speed.