Verified Biot-Savart using top grid
Switched to the cell-centered version of the vorticity so reduction works like with pressure which looks correct in the plots but not quantified yet.

The boundary reconstruction error (compared to the analytic solution) for the lamb vortex using the full vorticity field is 3e-6 when N=64 and reduces rapidly with more resolution.