Still working on a direct update to the residual
The projection loop is working but we are using p to sequentially update u, ω, u_BC, and σ just to update r each iteration. This has some issues:
1. This causes pressure to be driven to zero as we iterate, so we loose the pressure for the time step
2. Its a waste of compute to update all those intermediate fields.
3. It can't be integrated within `solve` directly.

It should be possible to go straight from p to ω to r, which would seem to solve everything. I've tried to code this up but it doesn't match the (working) indirect method yet.