Intermediate step
I've got the residual update matching the brute force approach, but only if I use ω from u. The `ω_from_p` isn't working yet.

To get this working, I needed to zero vorticity on the boundary and only compute the normal component of the velocity in `biotBC!`. It's also much more stable on increased resolution, so I increase to D=256.

The piece that I'm still fundamentally missing is why the boundary vorticity is so important. It goes to zero over the iterations, but convergence is much worse if it's excluded and faster if `boitBC!` is run repeatedly, assuring the boundary vorticity can influence the residual.