3D!
I've refactored the velocity function and the BC function to work in both 2D and 3D. Unlike most of WaterLily, this does require a little code duplication because in 2D you only need one `ω` component and the weight is 1/2πr^2, whereas in 3D you need all three components and the weight is  1/4πr^3. I used dispatch on the dimension of the vector and got the repeated code down to only 2 lines each (41&42, 72&73.

I also added a 3D vortex test case, a Hill vortex. It's actually nicer than the lamb dipole since it doesn't need special functions or autodiff, meaning it's probably the better case to roll into the final tests.

The new code works, but the accuracy sensitivity in 3D is quite different than 2D. Perhaps because the weights drop of cubically, but using `dis=1` reconstructs the Hill vortex without issues. Instead, the accuracy is very sensitive to the size of the lowest MLArray level, which needs to be around 8x8x8 for the Hill vortex. This drives the computational cost, with the rest of the refinement levels adding constant time. Perhaps this is due to placing the center of vorticity at the cell centers, which is the least accurate on the coarsest levels. Is the center computed using `|ω|` in 3D?

Between the larger number of boundary cells and the larger size of the lowest level, the 3D BC is pretty slow, ~0.5s on the Hill vortex with N=128. Hopefully, GPU to the rescue...