working group size fixed
Turns out that setting working group size 64 is the same as (64,1) in 2D and (64,1,1) in 3D. So slicing on i=1 and then using that working group means trying to split 1 thing 64 ways and not splitting the other dimensions at all!!! This was the REAL problem, with our GPU BC.

Fixed this by creating a tuple in @loop that assigns 64 to size(R)'s largest dimension. I'm sure this can be further optimized, but it keeps performance the same when using @inside, and makes all the i=1-slices much faster.

Tested and working in 2D and 3D vortices and the 2D circle case.