minimize u_ω calls
Refactored the projection step to minimize the number of times we have to call the (still somewhat expensive) u_ω function to only 1+nᵖ times per projection (down from 3+nᵖ). If nᵖ=1 after prediction and 0 after correction, this means 3 calls vs 7.
On the downside, if nᵖ>1, then p is overwritten and the velocity is (wastefully) updated more than once.
1. Compute ω from both u* AND old grad(p). Then we only call u_ω once on each domain face to find the initial residual.
2. If r₂<tol, update u with old grad(p). We already accounted for this, so no u_ω calls needed
3. If not, do a Vcycle, update u with new grad(p), and fix the domain faces with u_ω.
4. If still r₂>tol, reset p=0 and loop 3 to convergence.
5. Once outside the loop, update the ghost values, but not the domain faces.