Correct KernelAbstractions allocation
Kernel Abstractions was being passed a single workgroup value (64) but this was really setting the workgroup to (64,1 [,1]). In other words it was always splitting up the work along the first direction, even when passed a slice with only one element in that direction. Now it split into 64 groups along that largest axis.