Initial commit
This PR gets rid of the undocumented and slightly magical σᵥ divergence correction and replaces it with a documented and `residual!` function that corrects the residual at the start of every pressure `solver!` call.
This gets rid of the σᵥ field (reducing memory a bit) and should improve solver robustness at the expense of putting an `ifelse` statement @inside the residual calculation.