measure at timeNext instead of time
Change `sim_step!` to measure the body at timeNext instead of time This means the flow and BC are both at timeNext at the end of the time step. The results are exactly the same, just one time step earlier.
While cleaning up `sim_step!`, I also made it so calling the function without `t_end` does a single step, and added an accelerating body test case to check that the flow is synced to the body velocity.