AutoDiff time-step
The goal of this PR is to enable auto-diff with WaterLily.

A minor issue is that we have certain parts of the code which declare types which should not. I've removed some of these in the first commit.

The second and far more serious issue is that of the field arrays within `Flow` and `Poisson`. These are pre-allocated, and this is incompatible with ForwardDiff, which needs to be able to allocate Dual arrays for variables "downstream" of the derivative being taken. I've included an example "test.jl" file which fails with ForwardDiff when the simulation is pre-allocated as normal OR when pre-allocated with a Dual type.

The last issue is KernelAbstractions. It's possible that this will not be an issue once the types are sorted out, but the error messages are so bad that it's impossible to debug. I've changed `@loop` to just use @SIMD for the time being.