Updated working AD test
Got the very hacky proto-type test working. Checked it against a finite difference and it looks like it's correct!

This function uses non-in-place (out-of-place?) functions `AutoBody` and `measure` to generate the (nasty compound) Dual type needed for the derivative. Then the Flow is generated using that Type, and everything proceeds pretty much as normal from there.

This works because the example is taking the derivative with respect to a geometric parameter. If we were taking derivatives with respect to a flow parameter, I'm not sure how different the generated type would need to be and we would need out-of-place function calls to find out (which isn't generally how WaterLily is set up).