Update test.jl
I've cleaned up the step_force functions so I can come like-to-like for the mutating and non-mutating versions, with or without using Duals.

The two functions return identical value up to the 6 sig digit. I'm guessing this small error is from something which is not quite zeroed properly in mom_step, but I'm not sure.

The Dual version of the non-mutating function is weirdly sensitive to how things are initialized. When the function is first called with Dual types, it throws the `Cannot determine ordering of Dual tags` error, but if you re-evaluate the function and then call it again, it works. I don't know why yet.