Reinstate Periodic BCs
I changed the default `perdir=()` so that I cold test for an empty tuple in dispatch and loop `for i in perdir` otherwise. This is in a new `perBC!` function which is called within Poisson and after `solver!` is called in flow.

Unskipped the periodic tests and added a second allocation test for periodic flows. Passing and non-allocating.