Fix BCTuple
Evaluating BCTuple with t=t_0 in the predictor and t=t_0+dt in the corrector causes a mismatch between the BCs and the residuals (which integrate up to t_0+dt in both steps).
This makes the pressure solver extremely unhappy (especially on very large grids) and results in an incorrect pressure field.

This routine just sets U=Ui(t_0+dt) for the whole time step and adds a simple 2D test to make sure the solver recovers the potential flow solution for accelerating background flows efficiently. I've also tested this in 3D and on disks and it works consistently.